Modified chip system to avoid losing data when creating the shadow copies. All chip data operations are now in IChipData. The script InventoryChip was renamed to IChipDrag and only contains the dragging operations.
Chip data is updated when equipped and unequipped from character panel.
If a chip is equipped on load, the inventory displays a shadow copy of the corresponding equipped chip.

Equipped chips on load can be unequipped and return to their corresponding place in the inventory, instead of creating a duplicate.