Created script to force preload from any scene without having to change scenes manually.
Created a small test scene with only the player and an ending trigger. The scene is full of errors which will not be fixed.

Created a player manager to store money variable. In the future, it will also behave like "PlayerStats.cs".

Money is added when level is finished. There is no conversion.

Money is subtracted from player when buy objects.